Added a TestResult module and extract helper functions to internal package
