+++ artifact.go
@@ -4,16 +4,13 @@ import (
 	"encoding/xml"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"mime"
-	"mime/multipart"
 	"net/http"
 	"path/filepath"
-	"time"
 
 	"github.com/gin-gonic/gin"
+	"github.com/iver-wharf/wharf-api/internal/ctxparser"
 	"github.com/iver-wharf/wharf-core/pkg/ginutil"
-	"github.com/iver-wharf/wharf-core/pkg/problem"
 	"gorm.io/gorm"
 )
 
@@ -63,10 +60,6 @@ func (m artifactModule) Register(g *gin.RouterGroup) {
 	g.GET("/artifacts", m.getBuildArtifactsHandler)
 	g.GET("/artifact/:artifactId", m.getBuildArtifactHandler)
 	g.POST("/artifact", m.postBuildArtifactHandler)
-	g.POST("/test-result-data", m.postTestResultDataHandler)
-	g.GET("/test-result-details", m.getBuildAllTestResultDetailsHandler)
-	g.GET("/test-result-details/:artifactId", m.getBuildTestResultDetailsHandler)
-	g.GET("/test-results-summary", m.getBuildTestResultsSummaryHandler)
 	// deprecated
 	g.GET("/tests-results", m.getBuildTestsResultsHandler)
 }
@@ -113,7 +106,7 @@ func (m artifactModule) getBuildArtifactsHandler(c *gin.Context) {
 // @failure 502 {object} problem.Response "Database is unreachable"
 // @router /build/{buildid}/artifact/{artifactId} [get]
 func (m artifactModule) getBuildArtifactHandler(c *gin.Context) {
-	artifactID, buildID, ok := parseParamArtifactAndBuildID(c)
+	artifactID, buildID, ok := ctxparser.ParamArtifactAndBuildID(c)
 	if !ok {
 		return
 	}
@@ -144,6 +137,32 @@ func (m artifactModule) getBuildArtifactHandler(c *gin.Context) {
 
 	c.Header("Content-Disposition", disposition)
 	c.Data(http.StatusOK, mimeType, artifact.Data)
+}
+
+// postBuildArtifactHandler godoc
+// @summary Post build artifact
+// @tags artifact
+// @accept multipart/form-data
+// @param buildid path int true "Build ID"
+// @param file formData file true "Build artifact file"
+// @success 201 "Added new artifacts"
+// @failure 400 {object} problem.Response "Bad request"
+// @failure 401 {object} problem.Response "Unauthorized or missing jwt token"
+// @failure 404 {object} problem.Response "Artifact not found"
+// @failure 502 {object} problem.Response "Database is unreachable"
+// @router /build/{buildid}/artifact [post]
+func (m artifactModule) postBuildArtifactHandler(c *gin.Context) {
+	buildID, files, ok := ctxparser.ParamBuildIDAndFiles(c)
+	if !ok {
+		return
+	}
+
+	_, ok = createArtifacts(c, m.Database, files, buildID)
+	if !ok {
+		return
+	}
+
+	c.Status(http.StatusCreated)
 }
 
 // getBuildTestsResultsHandler godoc
@@ -196,386 +215,16 @@ func (m artifactModule) getBuildTestsResultsHandler(c *gin.Context) {
 	c.JSON(http.StatusOK, results)
 }
 
-// postBuildArtifactHandler godoc
-// @summary Post build artifact
-// @tags artifact
-// @accept multipart/form-data
-// @param buildid path int true "Build ID"
-// @param file formData file true "Build artifact file"
-// @success 201 "Added new artifacts"
-// @failure 400 {object} problem.Response "Bad request"
-// @failure 401 {object} problem.Response "Unauthorized or missing jwt token"
-// @failure 404 {object} problem.Response "Artifact not found"
-// @failure 502 {object} problem.Response "Database is unreachable"
-// @router /build/{buildid}/artifact [post]
-func (m artifactModule) postBuildArtifactHandler(c *gin.Context) {
-	buildID, files, ok := parseBuildIDAndFiles(c)
-	if !ok {
-		return
-	}
-
-	_, ok = m.createArtifacts(c, files, buildID)
-	if !ok {
-		return
-	}
-
-	c.Status(http.StatusCreated)
-}
-
-type file struct {
-	name     string
-	fileName string
-	data     []byte
-}
-
-// postTestResultDataHandler godoc
-// @summary Post test result data
-// @tags artifact
-// @accept multipart/form-data
-// @param buildid path int true "Build ID"
-// @param file formData file true "Test result artifact file"
-// @success 201 {object} TestResultListSummary "Added new test result data and created summary"
-// @failure 400 {object} problem.Response "Bad request"
-// @failure 502 {object} problem.Response "Database is unreachable"
-// @router /build/{buildid}/test-result-data [post]
-func (m artifactModule) postTestResultDataHandler(c *gin.Context) {
-	buildID, files, ok := parseBuildIDAndFiles(c)
-	if !ok {
-		return
-	}
-
-	artifacts, ok := m.createArtifacts(c, files, buildID)
-	if !ok {
-		return
-	}
-
-	summaries := make([]TestResultSummary, 0, len(artifacts))
-	lotsOfDetails := make([]TestResultDetail, 0)
-
-	listSummary := TestResultListSummary{
-		BuildID: buildID}
-	for _, artifact := range artifacts {
-		summary, details, err := getTestSummaryAndDetails(artifact.Data, artifact.ArtifactID, buildID)
-		if err != nil {
-			log.Warn().
-				WithError(err).
-				WithString("filename", artifact.Name).
-				WithUint("build", buildID).
-				WithUint("artifact", artifact.ArtifactID).
-				Message("Failed to unmarshal; invalid/unsupported TRX/XML format.")
-
-			ginutil.WriteProblemError(c, err,
-				problem.Response{
-					Type:   "/prob/api/test-results-parse",
-					Status: http.StatusBadGateway,
-					Title:  "Unexpected response format.",
-					Detail: fmt.Sprintf(
-						"Failed parsing test result ID %d, for build with ID %d in"+
-							" database. Invalid/unsupported TRX/XML format.", summary.ArtifactID, buildID),
-				})
-			return
-		}
-
-		listSummary.Failed += summary.Failed
-		listSummary.Passed += summary.Passed
-		listSummary.Skipped += summary.Skipped
-
-		summary.FileName = artifact.FileName
-		summaries = append(summaries, summary)
-		lotsOfDetails = append(lotsOfDetails, details...)
-	}
-
-	listSummary.Total = listSummary.Failed + listSummary.Passed + listSummary.Skipped
-
-	if err := m.Database.CreateInBatches(summaries, 10).Error; err != nil {
-		ginutil.WriteDBWriteError(c, err, fmt.Sprintf(
-			"Failed saving test result summaries for build with ID %d in database.",
-			buildID))
-		return
-	}
-
-	err := m.Database.
-		CreateInBatches(lotsOfDetails, 100).
-		Error
-	if err != nil {
-		ginutil.WriteDBWriteError(c, err, fmt.Sprintf(
-			"Failed saving test result details for build with ID %d in database.",
-			buildID))
-	}
-
-	c.JSON(http.StatusOK, listSummary)
-}
-
-// getBuildAllTestResultDetailsHandler godoc
-// @summary Get all test result details for specified build
-// @tags artifact
-// @param buildid path int true "Build ID"
-// @success 200 {object} []TestResultDetail
-// @failure 400 {object} problem.Response "Bad request"
-// @failure 502 {object} problem.Response "Database is unreachable"
-// @router /build/{buildid}/test-result-details [get]
-func (m artifactModule) getBuildAllTestResultDetailsHandler(c *gin.Context) {
-	buildID, ok := ginutil.ParseParamUint(c, "buildid")
-	if !ok {
-		return
-	}
-
-	details := []TestResultDetail{}
-	err := m.Database.
-		Where(&TestResultDetail{BuildID: buildID}).
-		Find(&details).
-		Error
-
-	if err != nil {
-		ginutil.WriteDBReadError(c, err, fmt.Sprintf(
-			"Failed fetching test result details for build with ID %d from database.",
-			buildID))
-		return
-	}
-
-	c.JSON(http.StatusOK, details)
-}
-
-// getBuildTestResultDetailsHandler godoc
-// @summary Get test result details for specified test
-// @tags artifact
-// @param buildid path int true "Build ID"
-// @param artifactId path int true "Artifact ID"
-// @success 200 {object} []TestResultDetail
-// @router /build/{buildid}/test-result-details/{artifactId} [get]
-func (m artifactModule) getBuildTestResultDetailsHandler(c *gin.Context) {
-	artifactID, buildID, ok := parseParamArtifactAndBuildID(c)
-	if !ok {
-		return
-	}
-
-	details := []TestResultDetail{}
-	err := m.Database.
-		Where(&TestResultDetail{BuildID: buildID, ArtifactID: artifactID}).
-		Find(&details).
-		Error
-
-	if err != nil {
-		ginutil.WriteDBReadError(c, err, fmt.Sprintf(
-			"Failed fetching test result details from test with ID %d for build with ID %d from database.",
-			artifactID, buildID))
-		return
-	}
-
-	c.JSON(http.StatusOK, details)
-}
-
-// getBuildTestResultsSummaryHandler godoc
-// @summary Get test result summary of all tests for specified build
-// @tags artifact
-// @param buildid path int true "Build ID"
-// @success 200 {object} TestResultListSummary
-// @failure 400 {object} problem.Response "Bad Request"
-// @failure 502 {object} problem.Response "Bad Gateway"
-// @router /build/{buildid}/test-results-summary [get]
-func (m artifactModule) getBuildTestResultsSummaryHandler(c *gin.Context) {
-	buildID, ok := ginutil.ParseParamUint(c, "buildid")
-	if !ok {
-		return
-	}
-
-	summaries := []TestResultSummary{}
-	err := m.Database.
-		Preload("Artifact").
-		Where(&TestResultSummary{BuildID: buildID}).
-		Find(&summaries).
-		Error
-
-	if err != nil {
-		ginutil.WriteDBReadError(c, err, fmt.Sprintf(
-			"Failed fetching test result summaries for build with ID %d from database.",
-			buildID))
-		return
-	}
-
-	listSummary := TestResultListSummary{
-		BuildID: buildID}
-
-	for _, v := range summaries {
-		listSummary.Failed += v.Failed
-		listSummary.Passed += v.Passed
-		listSummary.Skipped += v.Skipped
-	}
-
-	listSummary.Total = listSummary.Failed + listSummary.Passed + listSummary.Skipped
-
-	c.JSON(http.StatusOK, listSummary)
-}
-
-func parseMultipartFormData(c *gin.Context, buildID uint) ([]file, bool) {
-	form, err := c.MultipartForm()
-	if err != nil {
-		ginutil.WriteMultipartFormReadError(c, err,
-			fmt.Sprintf("Failed reading multipart-form content from request body when uploading new"+
-				" artifact for build with ID %d.", buildID))
-		return nil, false
-	}
-
-	var files []file
-	for k := range form.File {
-		if fhs := form.File[k]; len(fhs) > 0 {
-			data, ok := readMultipartFileData(c, buildID, fhs[0])
-			if !ok {
-				return nil, false
-			}
-
-			files = append(files, file{
-				name:     k,
-				fileName: fhs[0].Filename,
-				data:     data,
-			})
-		}
-	}
-
-	return files, true
-}
-
-func readMultipartFileData(c *gin.Context, buildID uint, fh *multipart.FileHeader) ([]byte, bool) {
-	if fh == nil {
-		return nil, false
-	}
-
-	f, err := fh.Open()
-	if err != nil {
-		ginutil.WriteMultipartFormReadError(c, err,
-			fmt.Sprintf("Failed with starting to read file content from multipart form request body when"+
-				" uploading new artifact for build with ID %d.", buildID))
-		return nil, false
-	}
-
-	defer func() {
-		if err := f.Close(); err != nil {
-			log.Error().
-				WithError(err).
-				Message("Failed to close multipart form request body file handle.")
-		}
-	}()
-
-	data, err := ioutil.ReadAll(f)
-	if err != nil {
-		ginutil.WriteMultipartFormReadError(c, err,
-			fmt.Sprintf("Failed reading file content from multipart form request body when uploading new"+
-				" artifact for build with ID %d.", buildID))
-		return nil, false
-	}
-
-	return data, true
-}
-
-type trxTestRun struct {
-	XMLName xml.Name `xml:"TestRun"`
-
-	Results struct {
-		XMLName         xml.Name `xml:"Results"`
-		UnitTestResults []struct {
-			XMLName   xml.Name `xml:"UnitTestResult"`
-			TestName  string   `xml:"testName,attr"`
-			Duration  string   `xml:"duration,attr"`
-			StartTime string   `xml:"startTime,attr"`
-			EndTime   string   `xml:"endTime,attr"`
-			Outcome   string   `xml:"outcome,attr"`
-			Output    struct {
-				XMLName   xml.Name `xml:"Output"`
-				ErrorInfo struct {
-					XMLName xml.Name `xml:"ErrorInfo"`
-					Message struct {
-						InnerXML string `xml:",innerxml"`
-					} `xml:"Message"`
-					StackTrace struct {
-						InnerXML string `xml:",innerxml"`
-					} `xml:"StackTrace"`
-				} `xml:"ErrorInfo"`
-			} `xml:"Output"`
-		} `xml:"UnitTestResult"`
-	} `xml:"Results"`
-
-	ResultSummary struct {
-		XMLName  xml.Name `xml:"ResultSummary"`
-		Counters struct {
-			XMLName             xml.Name `xml:"Counters"`
-			Total               uint     `xml:"total,attr"`
-			Executed            uint     `xml:"executed,attr"`
-			Passed              uint     `xml:"passed,attr"`
-			Failed              uint     `xml:"failed,attr"`
-			Error               uint     `xml:"error,attr"`
-			Timeout             uint     `xml:"timeout,attr"`
-			Aborted             uint     `xml:"aborted,attr"`
-			Inconclusive        uint     `xml:"inconclusive,attr"`
-			PassedButRunAborted uint     `xml:"passedButRunAborted,attr"`
-			NotRunnable         uint     `xml:"notRunnable,attr"`
-			NotExecuted         uint     `xml:"notExecuted,attr"`
-			Disconnected        uint     `xml:"disconnected,attr"`
-			Warning             uint     `xml:"warning,attr"`
-			Completed           uint     `xml:"completed,attr"`
-			InProgress          uint     `xml:"inProgress,attr"`
-			Pending             uint     `xml:"pending,attr"`
-		} `xml:"Counters"`
-	} `xml:"ResultSummary"`
-}
-
-// getTestSummaryAndDetails currently only supports the TRX/XML format.
-func getTestSummaryAndDetails(data []byte, artifactID, buildID uint) (TestResultSummary, []TestResultDetail, error) {
-	var testRun trxTestRun
-	if err := xml.Unmarshal(data, &testRun); err != nil {
-		return TestResultSummary{}, []TestResultDetail{}, err
-	}
-
-	details := make([]TestResultDetail, len(testRun.Results.UnitTestResults))
-	for idx, utr := range testRun.Results.UnitTestResults {
-		detail := &details[idx]
-		detail.ArtifactID = artifactID
-		detail.BuildID = buildID
-		detail.Name = utr.TestName
-		if utr.Outcome == "Passed" {
-			detail.Status = TestResultStatusSuccess
-		} else if utr.Outcome == "Failed" {
-			detail.Status = TestResultStatusFailed
-		} else if utr.Outcome == "NotExecuted" {
-			detail.Status = TestResultStatusSkipped
-		}
-		if detail.Status != TestResultStatusSuccess {
-			detail.Message.SetValid(fmt.Sprintf("%s\n%s",
-				utr.Output.ErrorInfo.Message.InnerXML,
-				utr.Output.ErrorInfo.StackTrace.InnerXML))
-		}
-
-		if startTime, err := time.Parse(time.RFC3339, utr.StartTime); err == nil {
-			detail.StartedOn.SetValid(startTime)
-		}
-
-		if endTime, err := time.Parse(time.RFC3339, utr.EndTime); err == nil {
-			detail.CompletedOn.SetValid(endTime)
-		}
-	}
-
-	counters := testRun.ResultSummary.Counters
-	summary := TestResultSummary{
-		ArtifactID: artifactID,
-		BuildID:    buildID,
-		Failed:     counters.Failed,
-		Passed:     counters.Passed,
-		Skipped:    counters.NotExecuted,
-		Total:      counters.Total,
-	}
-
-	return summary, details, nil
-}
-
-func (m artifactModule) createArtifacts(c *gin.Context, files []file, buildID uint) ([]Artifact, bool) {
+func createArtifacts(c *gin.Context, db *gorm.DB, files []ctxparser.File, buildID uint) ([]Artifact, bool) {
 	artifacts := make([]Artifact, len(files))
 	for idx, f := range files {
 		artifact := &artifacts[idx]
-		artifact.Data = f.data
-		artifact.Name = f.name
-		artifact.FileName = f.fileName
+		artifact.Data = f.Data
+		artifact.Name = f.Name
+		artifact.FileName = f.FileName
 		artifact.BuildID = buildID
 
-		err := m.Database.Create(artifact).Error
+		err := db.Create(artifact).Error
 		if err != nil {
 			ginutil.WriteDBWriteError(c, err, fmt.Sprintf(
 				"Failed saving artifact with name %q for build with ID %d in database.",
@@ -590,18 +239,4 @@ func (m artifactModule) createArtifacts(c *gin.Context, files []file, buildID ui
 			Message("File saved as artifact")
 	}
 	return artifacts, true
-}
-
-func parseParamArtifactAndBuildID(c *gin.Context) (artifactID, buildID uint, ok bool) {
-	if artifactID, ok = ginutil.ParseParamUint(c, "artifactId"); ok {
-		buildID, ok = ginutil.ParseParamUint(c, "buildid")
-	}
-	return
-}
-
-func parseBuildIDAndFiles(c *gin.Context) (buildID uint, files []file, ok bool) {
-	if buildID, ok = ginutil.ParseParamUint(c, "buildid"); ok {
-		files, ok = parseMultipartFormData(c, buildID)
-	}
-	return
 }
