+++ build.go
@@ -67,6 +67,9 @@ func (m buildModule) Register(g *gin.RouterGroup) {
 
 		artifacts := artifactModule{m.Database}
 		artifacts.Register(build)
+
+		testResults := testResultModule{m.Database}
+		testResults.Register(build)
 	}
 }
